+++ packages/xy/tests/scatter/Scatter.test.tsx
@@ -61,4 +61,26 @@ describe('Scatter', () => {
         // dataset has two series, so two colors
         expect(screen.queryAllByRole('legend-item')).toHaveLength(2)
     })
+
+    it('accepts data with keys', () => {
+        const p1 = { x: 1, y: 2, k: 5 }
+        const p2 = { x: 2, y: 4, k: 4 }
+        const p3 = { x: 4, y: 8, k: 3 }
+        const data = [{ id: 'A', data: [p1, p2, p3] }]
+        let processed: ProcessedDataContextProps = { data: [], seriesIndexes: {}, keys: [] }
+        const GetProcessedData = () => {
+            processed = useProcessedData()
+            return null
+        }
+        render(
+            <Chart>
+                <Scatter {...scatterProps} data={data} k={'k'}>
+                    <GetProcessedData />
+                </Scatter>
+            </Chart>
+        )
+        expect(Object.keys(processed.seriesIndexes)).toHaveLength(1)
+        expect(processed.data).toHaveLength(1)
+        expect(processed.data[0].k).toEqual([5, 4, 3])
+    })
 })