+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 
+-   [xy] props `k` in `Scatter` to provide explicit keys for data points
 -   [examples] chart with histogram with draggable controls
 -   [core] prop `disabledStyle` in legend components to control appearance of
     disabled legend items